Make libxml-ruby play nice with other consumers
`libxml-ruby` registers a global callback with libxml for node
deletions. If a node's `_private` field is not NULL, `libxml-ruby`
reaches through the pointer and clobbers a few fields. This interacts
badly with other consumers of the underlying libxml library, such as
Nokogiri.

Maintain a hashtable of node pointers which we own. Do not write through
any others.