Store libxml -> Ruby object mappings in a hashtable
Maintain a hashtable of mappings from libxml objects to Ruby objects by
address. This replaces the mechanism of using the _private pointer in
the libxml objects.

As libxml-ruby registers a global callback with libxml for node
deletions, when a node's _private field was not NULL, libxml-ruby
would reach through the pointer and clobber fields.
This interacts badly with other consumers of the underlying libxml
library, such as nokogiri, which use the _private field for their own
objects.

Addresses Github issue #116.